feat: Relocate Blend file to enable reuse of base color blending logic.
