+++ src/core/Blend.cpp
