+++ src/core/Blend.h
