+++ src/gpu/opengl/GLBlend.cpp
@@ -17,7 +17,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "GLBlend.h"
-#include "gpu/Blend.h"
+#include "core/Blend.h"
 
 namespace tgfx {
 static void HardLight(FragmentShaderBuilder* fsBuilder, const char* final, const char* src,
